+++ tests/observables/test_rx_amb.py
@@ -13,16 +13,17 @@ def test_rx_amb_init():
         rx_amb()
 
 
-def test_rx_amb(kernel):
+@pytest.mark.curio
+async def test_rx_amb():
 
     seeker = ObserverCounterCollector()
 
-    a = rx_create(subscribe=kernel.run(countdown(5, 0.1)))
-    b = rx_create(subscribe=kernel.run(countdown(10, 0.2)))
+    a = rx_create(subscribe=await countdown(5, 0.1))
+    b = rx_create(subscribe=await countdown(10, 0.2))
 
     obs = rx_amb(a, b)
-    sub_a = kernel.run(obs.subscribe(seeker))
-    kernel.run(sub_a())
+    sub_a = await obs.subscribe(seeker)
+    await sub_a()
 
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0
@@ -30,16 +31,17 @@ def test_rx_amb(kernel):
     assert seeker.items == [a]
 
 
-def test_rx_amb2(kernel):
+@pytest.mark.curio
+async def test_rx_amb2():
 
     seeker = ObserverCounterCollector()
 
-    a = rx_create(subscribe=kernel.run(countdown(5, 0.1)))
-    b = rx_create(subscribe=kernel.run(countdown(10, 0.05)))
+    a = rx_create(subscribe=await countdown(5, 0.1))
+    b = rx_create(subscribe=await countdown(10, 0.05))
 
     obs = rx_amb(a, b)
-    sub_a = kernel.run(obs.subscribe(seeker))
-    kernel.run(sub_a())
+    sub_a = await obs.subscribe(seeker)
+    await sub_a()
 
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0
@@ -47,7 +49,8 @@ def test_rx_amb2(kernel):
     assert seeker.items == [b]
 
 
-def test_rx_amb_with_error(kernel):
+@pytest.mark.curio
+async def test_rx_amb_with_error():
 
     seeker = ObserverCounterCollector()
 
@@ -56,11 +59,11 @@ def test_rx_amb_with_error(kernel):
         await an_observer.on_error(err="Args")
         return default_subscription
 
-    a = rx_create(subscribe=kernel.run(countdown(5, 0.1)))
+    a = rx_create(subscribe=await countdown(5, 0.1))
 
     obs = rx_amb(a, rx_create(subscribe=_subscribe))
-    sub_a = kernel.run(obs.subscribe(seeker))
-    kernel.run(sub_a())
+    sub_a = await obs.subscribe(seeker)
+    await sub_a()
 
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0