+++ tests/observables/test_rx_defer.py
@@ -6,7 +6,8 @@ from async_rx.protocol import default_subscription
 from ..model import ObserverCounter
 
 
-def test_rx_defer(kernel):
+@pytest.mark.curio
+async def test_rx_defer():
     async def _subscribe(an_observer: Observer):
         await an_observer.on_next(item=1)
         await an_observer.on_completed()
@@ -22,13 +23,13 @@ def test_rx_defer(kernel):
     assert seeker.on_next_count == 0
     assert seeker.on_completed_count == 0
     assert seeker.on_error_count == 0
-    unsub1 = kernel.run(obs.subscribe(seeker))
+    unsub1 = await obs.subscribe(seeker)
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0
 
     # even with max_observer=1 we can do that because of rx_defer
-    unsub2 = kernel.run(obs.subscribe(seeker))
+    unsub2 = await obs.subscribe(seeker)
     assert seeker.on_next_count == 2
     assert seeker.on_completed_count == 2
     assert seeker.on_error_count == 0