+++ tests/observables/test_rx_delay.py
@@ -19,29 +19,31 @@ def test_rx_delay_default():
         rx_delay(rx_from([1, 2]), duration=timedelta(seconds=0.5), buffer_size=-1)
 
 
-def test_rx_delay(kernel):
+@pytest.mark.curio
+async def test_rx_delay():
 
     source = rx_delay(rx_repeat_series([(0.1, "A"), (0.5, "B"), (1.0, "C")]), duration=timedelta(seconds=0.5), buffer_size=5)
 
     seeker = ObserverCounterCollectorWithTime()
 
-    sub = kernel.run(source.subscribe(seeker))
-    kernel.run(curio.sleep(3))
-    kernel.run(sub())
+    sub = await source.subscribe(seeker)
+    await curio.sleep(3)
+    await sub()
 
     assert len(seeker.items) == 3
     assert seeker.get_delta() == [0.5, 1.0]  # same interval all >= 0.5
 
 
-def test_rx_delay_with_error_should_delay_error(kernel):
+@pytest.mark.curio
+async def test_rx_delay_with_error_should_delay_error():
 
     source = rx_delay(rx_concat(rx_repeat_series([(0.1, "A"), (0.5, "B"), (1.0, "C")]), rx_throw("oups")), duration=timedelta(seconds=0.5))
 
     seeker = ObserverCounterCollectorWithTime()
 
-    sub = kernel.run(source.subscribe(seeker))
-    kernel.run(curio.sleep(3))
-    kernel.run(sub())
+    sub = await source.subscribe(seeker)
+    await curio.sleep(3)
+    await sub()
 
     assert len(seeker.items) == 3
     assert seeker.get_delta() == [0.5, 1.0]  # same interval all >= 0.5
@@ -50,15 +52,16 @@ def test_rx_delay_with_error_should_delay_error(kernel):
     assert seeker.on_completed_count == 0
 
 
-def test_rx_delay_minimal_interval(kernel):
+@pytest.mark.curio
+async def test_rx_delay_minimal_interval():
 
     source = rx_delay(rx_repeat_series([(0.1, "A"), (0.5, "B"), (1.0, "C")]), duration=timedelta(seconds=0.5))
 
     seeker = ObserverCounterCollectorWithTime()
 
-    sub = kernel.run(source.subscribe(seeker))
-    kernel.run(curio.sleep(3))
-    kernel.run(sub())
+    sub = await source.subscribe(seeker)
+    await curio.sleep(3)
+    await sub()
 
     assert len(seeker.items) == 3
     assert seeker.get_delta() == [0.5, 1.0]  # same interval because duration = 0.5
@@ -68,15 +71,16 @@ def test_rx_delay_minimal_interval(kernel):
 
     seeker = ObserverCounterCollectorWithTime()
 
-    sub = kernel.run(source.subscribe(seeker))
-    kernel.run(curio.sleep(5))
-    kernel.run(sub())
+    sub = await source.subscribe(seeker)
+    await curio.sleep(5)
+    await sub()
 
     assert len(seeker.items) == 3
     assert seeker.get_delta() == [1.5, 1.5]  # minimal duration is 1.5
 
 
-def test_rx_delay_overload_buffer(kernel):
+@pytest.mark.curio
+async def test_rx_delay_overload_buffer():
 
     _list = [(0.1, "A"), (0.1, "B"), (0.1, "C")] * 2
     assert len(_list) == 6
@@ -84,9 +88,9 @@ def test_rx_delay_overload_buffer(kernel):
     source = rx_delay(rx_repeat_series(_list), duration=timedelta(seconds=1.5), buffer_size=2, ignore_events_if_full=True)
     seeker = ObserverCounterCollectorWithTime()
 
-    sub = kernel.run(source.subscribe(seeker))
-    kernel.run(curio.sleep(10))
-    kernel.run(sub())
+    sub = await source.subscribe(seeker)
+    await curio.sleep(10)
+    await sub()
 
     # rx_repeat_series => O.6 second,
     # A, B, C because buffer size = 2, other are ignored