+++ tests/observables/test_rx_distinct.py
@@ -8,11 +8,12 @@ from ..model import ObserverCounterCollector
 from .model import get_observable
 
 
-def test_rx_distinct_dummy(kernel):
+@pytest.mark.curio
+async def test_rx_distinct_dummy():
     seeker = ObserverCounterCollector()
 
-    sub = kernel.run(rx_distinct(observable=get_observable(), frame_size=3).subscribe(an_observer=seeker))
-    kernel.run(sub())
+    sub = await rx_distinct(observable=get_observable(), frame_size=3).subscribe(an_observer=seeker)
+    await sub()
 
     assert seeker.on_next_count == 100
     assert seeker.on_completed_count == 1
@@ -26,13 +27,14 @@ def test_rx_distinct_dummy(kernel):
         rx_distinct(observable=get_observable(), frame_size=-1)
 
 
-def test_rx_distinct(kernel):
+@pytest.mark.curio
+async def test_rx_distinct():
 
     source = lambda: rx_from(observable_input=[1, 2, 1, 3, 2, 1, 4, 3, 2, 1, 5, 4, 3, 2, 1])
 
     seeker = ObserverCounterCollector()
-    sub = kernel.run(rx_distinct(observable=source(), frame_size=3).subscribe(an_observer=seeker))
-    kernel.run(sub())
+    sub = await rx_distinct(observable=source(), frame_size=3).subscribe(an_observer=seeker)
+    await sub()
 
     assert seeker.on_next_count == 9
     assert seeker.on_completed_count == 1
@@ -40,8 +42,8 @@ def test_rx_distinct(kernel):
     assert seeker.items == [1, 2, 3, 4, 1, 5, 3, 2, 1]
 
     seeker = ObserverCounterCollector()
-    sub = kernel.run(rx_distinct(observable=source(), frame_size=2).subscribe(an_observer=seeker))
-    kernel.run(sub())
+    sub = await rx_distinct(observable=source(), frame_size=2).subscribe(an_observer=seeker)
+    await sub()
     assert seeker.on_next_count == 13
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0