+++ tests/observables/test_rx_map.py
@@ -1,17 +1,20 @@
+import pytest
+
 from async_rx import rx_from, rx_map
 
 from ..model import ObserverCounterCollector
 
 
-def test_rx_map_with_async(kernel):
+@pytest.mark.curio
+async def test_rx_map_with_async():
 
     seeker = ObserverCounterCollector()
 
     async def add(i):
         return i + 2
 
-    sub_a = kernel.run(rx_map(observable=rx_from([1, 2, 3]), transform=add).subscribe(seeker))
-    kernel.run(sub_a())
+    sub_a = await rx_map(observable=rx_from([1, 2, 3]), transform=add).subscribe(seeker)
+    await sub_a()
 
     assert seeker.on_completed_count == 1
     assert seeker.on_next_count == 3
@@ -19,15 +22,16 @@ def test_rx_map_with_async(kernel):
     assert seeker.items == [3, 4, 5]
 
 
-def test_rx_map_sync(kernel):
+@pytest.mark.curio
+async def test_rx_map_sync():
 
     seeker = ObserverCounterCollector()
 
     def add(i):
         return i + 3
 
-    sub_a = kernel.run(rx_map(observable=rx_from([1, 2, 3]), transform=add).subscribe(seeker))
-    kernel.run(sub_a())
+    sub_a = await rx_map(observable=rx_from([1, 2, 3]), transform=add).subscribe(seeker)
+    await sub_a()
 
     assert seeker.on_completed_count == 1
     assert seeker.on_next_count == 3
@@ -35,15 +39,16 @@ def test_rx_map_sync(kernel):
     assert seeker.items == [4, 5, 6]
 
 
-def test_rx_map_with_expand_args(kernel):
+@pytest.mark.curio
+async def test_rx_map_with_expand_args():
 
     seeker = ObserverCounterCollector()
 
     def add(i, j):
         return i + j
 
-    sub_a = kernel.run(rx_map(observable=rx_from([(1, 2), (3, 4), (5, 6)]), transform=add, expand_arg_parameters=True).subscribe(seeker))
-    kernel.run(sub_a())
+    sub_a = await rx_map(observable=rx_from([(1, 2), (3, 4), (5, 6)]), transform=add, expand_arg_parameters=True).subscribe(seeker)
+    await sub_a()
 
     assert seeker.on_completed_count == 1
     assert seeker.on_next_count == 3
@@ -51,15 +56,16 @@ def test_rx_map_with_expand_args(kernel):
     assert seeker.items == [3, 7, 11]
 
 
-def test_rx_map_with_expand_kwargs(kernel):
+@pytest.mark.curio
+async def test_rx_map_with_expand_kwargs():
 
     seeker = ObserverCounterCollector()
 
     def inc(i):
         return i + 1
 
-    sub_a = kernel.run(rx_map(observable=rx_from([{'i': 1}, {'i': 4}]), transform=inc, expand_kwarg_parameters=True).subscribe(seeker))
-    kernel.run(sub_a())
+    sub_a = await rx_map(observable=rx_from([{'i': 1}, {'i': 4}]), transform=inc, expand_kwarg_parameters=True).subscribe(seeker)
+    await sub_a()
 
     assert seeker.on_completed_count == 1
     assert seeker.on_next_count == 2