+++ tests/observables/test_rx_of.py
@@ -1,15 +1,18 @@
+import pytest
+
 from async_rx import Observable, Observer, rx_of
 
 from ..model import ObserverCounterCollector
 
 
-def test_rx_of(kernel):
+@pytest.mark.curio
+async def test_rx_of():
 
     obs: Observable = rx_of(1, 2, 3)
 
     seeker = ObserverCounterCollector()
 
-    kernel.run(obs.subscribe(seeker))
+    await obs.subscribe(seeker)
 
     assert seeker.on_next_count == 3
     assert seeker.on_completed_count == 1