+++ tests/observables/test_rx_range.py
@@ -1,15 +1,18 @@
+import pytest
+
 from async_rx import Observable, Observer, rx_range
 from async_rx.protocol import default_subscription
 
 from ..model import ObserverCounterCollector
 
 
-def test_rx_range(kernel):
+@pytest.mark.curio
+async def test_rx_range():
     obs: Observable = rx_range(start=1, stop=4)
 
     seeker = ObserverCounterCollector()
 
-    kernel.run(obs.subscribe(seeker))
+    await obs.subscribe(seeker)
 
     assert seeker.on_next_count == 3
     assert seeker.on_completed_count == 1