+++ tests/observables/test_rx_throttle.py
@@ -17,7 +17,8 @@ def test_rx_throttle_default():
         rx_throttle(None, lambda a: not a)
 
 
-def test_rx_throttle(kernel):
+@pytest.mark.curio
+async def test_rx_throttle():
     source = rx_throttle(
         rx_repeat_series([(0.1, "A0"), (0.1, "A1"), (0.1, "A2"), (0.1, "A3"), (0.5, "B"), (1.0, "C"), (3.0, "D")]), duration=timedelta(seconds=0.5)
     )
@@ -32,9 +33,9 @@ def test_rx_throttle(kernel):
     #     5.3 -> complete
     seeker = ObserverCounterCollectorWithTime()
 
-    sub = kernel.run(source.subscribe(seeker))
-    kernel.run(curio.sleep(6))
-    kernel.run(sub())
+    sub = await source.subscribe(seeker)
+    await curio.sleep(6)
+    await sub()
     assert seeker.on_next_count == 4
     assert seeker.on_error_count == 0
     assert seeker.on_completed_count == 1
@@ -47,7 +48,8 @@ def test_rx_throttle(kernel):
     assert seeker.get_delta() == [0.8, 1.0, 3.0]  # see duration
 
 
-def test_rx_throttle_with_error(kernel):
+@pytest.mark.curio
+async def test_rx_throttle_with_error():
     source = rx_throttle(
         rx_concat(rx_repeat_series([(0.1, "A0"), (0.1, "A1"), (0.1, "A2"), (0.1, "A3"), (0.5, "B"), (1.0, "C"), (3.0, "D")]), rx_throw("oups")),
         duration=timedelta(seconds=0.5),
@@ -55,9 +57,9 @@ def test_rx_throttle_with_error(kernel):
 
     seeker = ObserverCounterCollectorWithTime()
 
-    sub = kernel.run(source.subscribe(seeker))
-    kernel.run(curio.sleep(6))
-    kernel.run(sub())
+    sub = await source.subscribe(seeker)
+    await curio.sleep(6)
+    await sub()
     assert seeker.on_next_count == 4
     assert seeker.on_error_count == 1
     assert seeker.on_completed_count == 0