+++ tests/observables/test_rx_throw.py
@@ -1,14 +1,17 @@
+import pytest
+
 from async_rx import Observable, Observer, rx_throw
 
 from ..model import ObserverCounterCollector
 
 
-def test_rx_throw(kernel):
+@pytest.mark.curio
+async def test_rx_throw():
     obs: Observable = rx_throw(error="oulala")
 
     seeker = ObserverCounterCollector()
 
-    kernel.run(obs.subscribe(seeker))
+    await obs.subscribe(seeker)
 
     assert seeker.on_next_count == 0
     assert seeker.on_completed_count == 0