+++ tests/protocol/test_observer.py
@@ -3,11 +3,13 @@ import pytest
 from async_rx.protocol import default_error
 
 
-def test_default_error_with_message(kernel):
+@pytest.mark.curio
+async def test_default_error_with_message(kernel):
     with pytest.raises(Exception):
-        kernel.run(default_error(err="big hup"))
+        await default_error(err="big hup")
 
 
-def test_default_error_with_Exception(kernel):
+@pytest.mark.curio
+async def test_default_error_with_Exception(kernel):
     with pytest.raises(Exception):
-        kernel.run(default_error(err=RuntimeError("Ouch")))
+        await default_error(err=RuntimeError("Ouch"))